requirements.txt
einops
langchain-groq
sentence_transformers
faiss-cpu
langchain-community
langchain
pypdf
huggingface_hub